+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 # Release Content
 ## Additions
 
-[Take into account the namespace for the attach hostedcluster #205](https://github.com/stolostron/cm-cli/issues/205)
-
+- [Take into account the namespace for the attach hostedcluster #205](https://github.com/stolostron/cm-cli/issues/205)
+- [Add cmd to run/hibernate/skip schedule hive cluster #207](https://github.com/stolostron/cm-cli/issues/207)
+    - `cm hibernate cluster <cluster_name>[,<clusterc_name>...] <options>`
+    - `cm run cluster <cluster_name>[,<clusterc_name>...] <options>`
+    - `cm set cluster <cluster_name>[,<clusterc_name>...] [--hibernate-schedule-on|--hibernate-schedule-off]`
 ## Breaking changes
 
 ## Bug fixes